input investmenttype on import
