+++ src/components/Modal/ImportTransactionsModalContent.js
@@ -1,0 +1,125 @@
+import { Button, Input, ModalBody, ModalContent, ModalFooter, ModalHeader } from "@nextui-org/react"
+import { useRef, useState } from "react";
+import { useTranslation } from "react-i18next"
+import { useDispatch } from "react-redux";
+import { UploadIcon } from "../../icons/UploadIcon";
+import { textInputStyle } from "../../Util/Global";
+import { importTransactions } from "../../actions/accounts";
+
+export default function ImportTransactionsModalContent({ accountKey }) {
+
+    const { t } = useTranslation()
+
+    const dispatch = useDispatch();
+    const [investmentType, setInvestmentType] = useState("")
+
+    const hiddenFileInput = useRef(null);
+
+    const handleClick = () => {
+        hiddenFileInput.current.click();
+    };
+
+    function getHoldings(accountKey, transactions, type) {
+        console.log(transactions)
+        if (!transactions) {
+            return
+        }
+        const holdings = []
+        const uniqueHoldingKeys = [...new Set(transactions.map(transaction => transaction.e24Key))];
+        console.log(uniqueHoldingKeys)
+        uniqueHoldingKeys.forEach(e24Key => {
+            const equityShare = transactions.filter(transaction => transaction.e24Key === e24Key).reduce((sum, transaction) => sum + parseFloat(transaction.equityShare), 0)
+
+            if (equityShare > 0) {
+                holdings.push(
+                    {
+                        name: transactions.find(transaction => transaction.e24Key === e24Key).name,
+                        accountKey: accountKey,
+                        type: "",
+                        equityShare,
+                        equityType: type,
+                        e24Key,
+                        goalPercentage: 0
+                    }
+                )
+            }
+        })
+        console.log(holdings)
+        return holdings
+    }
+
+    function readCsv(event) {
+        var transactions = [];
+        if (event.target.files && event.target.files[0]) {
+            const input = event.target.files[0];
+            const reader = new FileReader();
+            reader.onload = function (event) {
+                const file = event.target.result.split('\n');
+                file.forEach((line, index) => {
+                    if (index !== 0) {
+                        const data = line.split(',');
+                        transactions.push({
+                            key: data[0],
+                            cost: parseFloat(data[1]),
+                            name: data[2],
+                            type: data[3],
+                            date: data[4],
+                            equityPrice: parseFloat(data[5]),
+                            e24Key: data[6],
+                            equityShare: parseFloat(data[7])
+                        });
+                    }
+                })
+
+                const holdings = getHoldings(accountKey, transactions, investmentType)
+                dispatch(importTransactions({ key: accountKey, transactions, holdings }))
+            };
+            reader.readAsText(input);
+        }
+    };
+
+    return (
+        <ModalContent>
+            {(onClose) => (
+                <>
+                    <ModalHeader className="justify-between">
+                        <div className="flex gap-5">
+                            <div className="flex flex-col gap-1 items-start justify-center">
+                                <h4 className="text-small font-semibold leading-none text-default-600">Importer transaksjoner</h4>
+                            </div>
+                        </div>
+
+                    </ModalHeader>
+                    <ModalBody>
+
+                        <div className='grid grid-cols-2 gap-12 items-center justify-between'>
+                            <Input type="text"
+                                classNames={textInputStyle}
+                                label={"Investeringstype"}
+                                value={investmentType}
+                                onValueChange={setInvestmentType} />
+                            <Button color="primary" variant="bordered" onPress={handleClick} size="lg">
+                                <input type="file"
+                                    ref={hiddenFileInput}
+                                    onChange={readCsv}
+                                    accept=".csv"
+                                    style={{ display: 'none' }} />
+                                Last opp fil <UploadIcon />
+                            </Button>
+                        </div>
+                    </ModalBody>
+                    <ModalFooter>
+                        <Button variant="light" onPress={() => {
+                            onClose()
+                        }}>
+                            Lukk
+                        </Button>
+                        <Button color="primary" variant="light" onPress={onClose}>
+                            {t('investmentModal.close')}
+                        </Button>
+                    </ModalFooter>
+                </>
+            )}
+        </ModalContent>
+    )
+}
