+++ src/components/TransactionsTable.js
@@ -1,9 +1,11 @@
 import { useEffect, useMemo, useRef, useState } from "react";
-import { Button, Spacer, Table, TableBody, TableCell, TableColumn, TableHeader, TableRow, getKeyValue } from "@nextui-org/react";
+import { Button, Spacer, Table, TableBody, TableCell, TableColumn, TableHeader, TableRow, getKeyValue, useDisclosure } from "@nextui-org/react";
 import { calculateValue, getTransactionsFromFiri, getValueInFiat } from "../Util/Firi";
 import { useDispatch } from "react-redux";
 import { importTransactions } from "../actions/accounts";
 import { UploadIcon } from "../icons/UploadIcon";
+import EmptyModal from "./Modal/EmptyModal";
+import ImportTransactionsModalContent from "./Modal/ImportTransactionsModalContent";
 
 export default function TransactionsTable({ account }) {
 
@@ -13,6 +15,8 @@ export default function TransactionsTable({ account }) {
     });
 
     const dispatch = useDispatch()
+
+    const { onOpen, isOpen, onOpenChange } = useDisclosure();
 
     const sortedItems = useMemo(() => {
         return [...account.transactions].sort((a, b) => {
@@ -39,7 +43,7 @@ export default function TransactionsTable({ account }) {
     ];
 
     useEffect(() => {
-        if (account.type !== 'Automatic') {
+        if (account.type !== 'Cryptocurrency') {
             return
         }
 
@@ -114,81 +118,17 @@ export default function TransactionsTable({ account }) {
         fetchData()
     }, [])// eslint-disable-line react-hooks/exhaustive-deps
 
-    const hiddenFileInput = useRef(null);
-
-    const handleClick = () => {
-        hiddenFileInput.current.click();
-    };
-
-    function getHoldings(accountKey, transactions, type) {
-        console.log(transactions)
-        if (!transactions) {
-            return
-        }
-        const holdings = []
-        const uniqueHoldingKeys = [...new Set(transactions.map(transaction => transaction.e24Key))];
-        console.log(uniqueHoldingKeys)
-        uniqueHoldingKeys.forEach(e24Key => {
-            const equityShare = transactions.filter(transaction => transaction.e24Key === e24Key).reduce((sum, transaction) => sum + parseFloat(transaction.equityShare), 0)
-
-            if (equityShare > 0) {
-                holdings.push(
-                    {
-                        name: transactions.find(transaction => transaction.e24Key === e24Key).name,
-                        accountKey: accountKey,
-                        equityShare,
-                        equityType: type,
-                        e24Key,
-                        goalPercentage: 0
-                    }
-                )
-            }
-        })
-        console.log(holdings)
-        return holdings
-    }
-
-    function readCsv(event) {
-        var transactions = [];
-        if (event.target.files && event.target.files[0]) {
-            const input = event.target.files[0];
-            const reader = new FileReader();
-            reader.onload = function (event) {
-                const file = event.target.result.split('\n');
-                file.forEach((line, index) => {
-                    if (index !== 0) {
-                        const data = line.split(',');
-                        transactions.push({
-                            key: data[0],
-                            cost: parseFloat(data[1]),
-                            name: data[2],
-                            type: data[3],
-                            date: data[4],
-                            equityPrice: parseFloat(data[5]),
-                            e24Key: data[6],
-                            equityShare: parseFloat(data[7])
-                        });
-                    }
-                })
-
-                const holdings = getHoldings(account.key, transactions, account.type)
-                dispatch(importTransactions({ key: account.key, transactions, holdings }))
-            };
-            reader.readAsText(input);
-        }
-    };
-
     return (
         <div>
-            {account.type === 'Automatic' ? <></> :
-                <Button color="primary" variant="bordered" onPress={handleClick} size="lg">
-                    <input type="file"
-                        ref={hiddenFileInput}
-                        onChange={readCsv}
-                        accept=".csv"
-                        style={{ display: 'none' }} />
-                    Importer transaksjoner <UploadIcon />
-                </Button>
+            {account.type === 'Cryptocurrency' ? <></> :
+                <>
+                    <EmptyModal isOpen={isOpen} onOpenChange={onOpenChange} hideCloseButton={false} isDismissable={true}>
+                        <ImportTransactionsModalContent accountKey={account.key} />
+                    </EmptyModal>
+                    <Button color="primary" variant="bordered" onPress={onOpen} size="lg">
+                        Importer transaksjoner <UploadIcon />
+                    </Button>
+                </>
             }
 
             <Spacer y={4} />
